Changed graphics to use namespaces
